requirements/optional.txt
-r base.txt
djangorestframework>=3.3
python-jose>=0.6.1